chatbot/Chatbot/requirements.txt
google-generativeai
requests
numexpr